update_playcount and update_last_played moved to playlists.py
